requirement.txt

mysql.connector
secure-smtplib==0.1.1
pandas
lxml